docs: move remote content files into subdirectory
