+++ website/docusaurus.config.js
@@ -1,5 +1,5 @@
-const { readmes } = require('./readme-fns');
-const { externalDocs } = require('./external-docs');
+const { sdks } = require('./remote-content/sdks');
+const { externalDocs } = require('./remote-content/edge-proxy');
 
 // for a given redirect object, modify it's `from` property such that for every
 // path that doesn't start with `/docs/`, a corresponding path that _does_ start
@@ -614,8 +614,8 @@ module.exports = {
                 name: 'content-sdks',
                 sourceBaseUrl: 'https://raw.githubusercontent.com/Unleash/', // gets prepended to all of the documents when fetching
                 outDir: 'docs/generated', // the base directory to output to.
-                documents: readmes.urls, // the file names to download
-                modifyContent: readmes.modifyContent,
+                documents: sdks.urls, // the file names to download
+                modifyContent: sdks.modifyContent,
             },
         ],
         [
